tried to protect all Eigen constructors from complaining about wrong argument type when used with spline.getDimension, getPointSize or getSplineOrder by an explicit cast to int.

+++ aslam_splines/test/TestOPTBSpline.cpp
@@ -110,8 +110,8 @@ struct OPTSplineTester{
 		TestBSpline bsplineCopyConstruct(bspline), bsplineCopyAssign(createMyConf());
 		bsplineCopyAssign = bspline;
 
-		SM_ASSERT_EQ(std::runtime_error, ISplineOrder, bspline.getSplineOrder(), "");
-		SM_ASSERT_EQ(std::runtime_error, IDim, bspline.getDimension(), "");
+		SM_ASSERT_EQ(std::runtime_error, ISplineOrder, (int)bspline.getSplineOrder(), "");
+		SM_ASSERT_EQ(std::runtime_error, IDim, (int)bspline.getDimension(), "");
 	}
 
 	void testUpdate(){
@@ -242,7 +242,7 @@ struct OPTSplineTester{
 
 		for(int i = 0; i < 10; i ++){
 			Eigen::MatrixXd startPoint = manifold.getRandomPoint();
-			Eigen::VectorXd updateVector = Eigen::VectorXd::Random(manifold.getDimension());
+			Eigen::VectorXd updateVector = Eigen::VectorXd::Random((int)manifold.getDimension());
 
 			dv.setParameters(startPoint);
 			dv.update(&updateVector[0], updateVector.size());
